# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teacher/classes:
    get:
      operationId: getTeacherClasses
      tags: [teacher]
      summary: List the teacher's classes with their students
      description: Admins can pass ?teacherId= to preview another teacher's dashboard.
      parameters:
        - $ref: "#/components/parameters/TeacherIdQuery"
      responses:
        "200":
          description: Classes list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherClassList"
        "401":
          description: Not signed in
        "403":
          description: Not a teacher or admin
  /teacher/unclaimed:
    get:
      operationId: getTeacherUnclaimed
      tags: [teacher]
      summary: List students who are not assigned to any class yet
      parameters:
        - $ref: "#/components/parameters/TeacherIdQuery"
      responses:
        "200":
          description: Unclaimed students list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnclaimedStudentList"
        "401":
          description: Not signed in
        "403":
          description: Not a teacher or admin
...
  /teacher/students/{id}/claim:
    post:
      operationId: claimTeacherStudent
      tags: [teacher]
      summary: Add an unclaimed student to one of the teacher's classes
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - $ref: "#/components/parameters/TeacherIdQuery"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimStudentBody"
      responses:
        "200":
          description: Student claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherStudent"
        "401":
          description: Not signed in
        "403":
          description: Not a teacher or admin
  /teacher/students/{id}:
    patch:
      operationId: updateTeacherStudent
      tags: [teacher]
      summary: Update a student's name or points in the teacher's class
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - $ref: "#/components/parameters/TeacherIdQuery"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeacherStudentBody"
      responses:
        "200":
          description: Student updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherStudent"
        "401":
          description: Not signed in
        "403":
          description: Not a teacher or admin
  /teacher/students/{id}/class:
    delete:
      operationId: removeTeacherStudentClass
      tags: [teacher]
      summary: Remove a student from the teacher's class
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - $ref: "#/components/parameters/TeacherIdQuery"
      responses:
        "200":
          description: Student removed from class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherStudentClass"
        "401":
          description: Not signed in
        "403":
          description: Not a teacher or admin
...
components:
  parameters:
...
    TeacherIdQuery:
      in: query
      name: teacherId
      required: false
      schema:
        type: integer
      description: Optional teacher id (admin-only) to preview another teacher's dashboard.
  schemas:
...
    AvatarConfig:
      type: object
      properties:
        bgColor:
          type: string
        accessories:
          type: array
          items:
            type: string
        gender:
          type: string
          enum: [male, female]
        pet:
          type: string
        nickname:
          type: string
          maxLength: 30
          default: ""
        frame:
          type: string
          default: "none"
        badges:
          type: array
          items:
            type: string
          default: []
      required:
        - bgColor
        - accessories
        - gender
        - pet
        - nickname
        - frame
        - badges
...
    TeacherClass:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        teacherId:
          oneOf:
            - type: integer
            - type: "null"
        isChatEnabled:
          type: boolean
        students:
          type: array
          items:
            $ref: "#/components/schemas/TeacherStudent"
      required:
        - id
        - name
        - teacherId
        - isChatEnabled
        - students
    TeacherClassList:
      type: object
      properties:
        classes:
          type: array
          items:
            $ref: "#/components/schemas/TeacherClass"
      required:
        - classes
    TeacherStudent:
      type: object
      properties:
        id:
          type: integer
        replitUserId:
          oneOf:
            - type: string
            - type: "null"
        name:
          type: string
        email:
          oneOf:
            - type: string
            - type: "null"
        points:
          type: integer
        avatarConfig:
          $ref: "#/components/schemas/AvatarConfig"
      required:
        - id
        - replitUserId
        - name
        - points
        - avatarConfig
    UnclaimedStudentList:
      type: object
      properties:
        students:
          type: array
          items:
            $ref: "#/components/schemas/TeacherStudent"
      required:
        - students
    ClaimStudentBody:
      type: object
      properties:
        classId:
          type: integer
      required:
        - classId
    UpdateTeacherStudentBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        points:
          type: integer
      required: []
    TeacherStudentClass:
      type: object
      properties:
        id:
          type: integer
        classId:
          oneOf:
            - type: integer
            - type: "null"
      required:
        - id
        - classId
...